Fix Virtual Camera movement mapping
Swap positive and negative values for angular.y and angular.z to fit results from testing on a Bebop2. Virtual camera appears to have range [-83,17] for angular.y and range [-35,35] for angular.z. Note that this does not fit the description above so perhaps it has changed between Bebop 1 and 2 in which case the + direction may have also changed. Please verify.